gpt_telegram_bot.py
import os
import openai
from telegram import Update
from telegram.ext import ApplicationBuilder, MessageHandler, ContextTypes, filters

TELEGRAM_TOKEN = os.getenv("TELEGRAM_TOKEN")
OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")

openai.api_key = OPENAI_API_KEY

async def handle_message(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
    user_text = update.message.text
    try:
        resp = openai.ChatCompletion.create(
            model="gpt-3.5-turbo",
            messages=[{"role": "user", "content": user_text}],
        )
        reply = resp.choices[0].message.content.strip()
    except Exception as e:
        reply = f"Error: {e}"
    await update.message.reply_text(reply)

def main() -> None:
    if not TELEGRAM_TOKEN or not OPENAI_API_KEY:
        raise EnvironmentError("TELEGRAM_TOKEN and OPENAI_API_KEY must be set")
    app = ApplicationBuilder().token(TELEGRAM_TOKEN).build()
    app.add_handler(MessageHandler(filters.TEXT & ~filters.COMMAND, handle_message))
    app.run_polling()

if __name__ == "__main__":
    main()
